requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2022.5.18.1
    # via requests
chardet==4.0.0
    # via requests
click==8.1.3
    # via typer
colorama==0.4.4
    # via click
greenlet==1.1.2
    # via playwright
idna==2.10
    # via requests
playwright==1.22.0
    # via -r requirements.in
pyee==8.1.0
    # via playwright
python-slugify==4.0.1
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
text-unidecode==1.3
    # via python-slugify
typer==0.4.1
    # via -r requirements.in
urllib3==1.26.9
    # via requests
websockets==10.1
    # via playwright
